pyproject.toml
[project]
name = "beets-mpd-utils"
version = "0.0.1"
description = "Beets plugins to interface with MPD"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
keywords = ["beets", "MPD", "music"]

dependencies = [
  "beets",
  "python-mpd2",
]

[project.urls]
Homepage = "https://github.com/thekakkun/beets-mpd-utils"
Repository = "https://github.com/thekakkun/beets-mpd-utils.git"

[tool.pylint.typecheck]
generated-members = [
  "beetsplug.mpd_dj.MPDQueue.add",
  "beetsplug.mpd_dj.MPDQueue.playlist",
  "beetsplug.mpd_dj.MPDQueue.random",
  "beetsplug.mpd_dj.MPDQueue.status",
  "beetsplug.tracker.MPDTracker.currentsong",
  "beetsplug.tracker.MPDTracker.status",
]

[tool.setuptools.packages.find]
where = ["src"]
